Tests/LibWeb/Text/expected/css-has-invalidation/has-control-no-has-with-text-insert.txt

[after text insert (no :has() rule)]
  hasAncestorWalkInvocations: 0
  hasAncestorWalkVisits: 0
  hasMatchInvocations: 0
  hasResultCacheHits: 0
  hasResultCacheMisses: 0
  styleInvalidations: 2
